dehydrate multichain test mocks and utilities for fixture resuse in multiple suites.
